requirements.txt
amqp==5.0.6
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
autopep8==1.6.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.9
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
flake8==4.0.1
gunicorn==20.1.0
h11==0.12.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.2
kombu==5.2.0
lxml==4.6.4
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
outcome==1.1.0
Pillow==8.4.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-dotenv==0.19.1
python3-openid==3.2.0
pytz==2021.3
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sortedcontainers==2.4.0
soupsieve==2.3
sqlparse==0.4.2
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wsproto==1.0.0
